Extract anonymous class out into CheckExplainConfig
